PXF-JDBC: Make 'DbProduct' an enum
Make 'DbProduct' an enum, so that its implementations
did not have to be constructed every time 'getDbProduct()' is called.

This removes all subclasses of 'DbProduct'. Their code moves inside
'DbProduct' itself.

Proposed by Alexander Denissov <adenissov@pivotal.io>
