--- server/pxf-jdbc/src/main/java/org/greenplum/pxf/plugins/jdbc/utils/OracleProduct.java
@@ -1,35 +1,0 @@
-package org.greenplum.pxf.plugins.jdbc.utils;
-
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-/**
- * Implements methods for Oracle Database.
- */
-public class OracleProduct extends DbProduct {
-    @Override
-    public String wrapDate(Object val) {
-        return "to_date('" + val + "', 'YYYY-MM-DD')";
-    }
-
-    @Override
-    public String wrapTimestamp(Object val) {
-        return "to_timestamp('" + val + "', 'YYYY-MM-DD HH:MI:SS.FF')";
-    }
-}
